// src/docs/pbes/node/tag_pb.ts
// @generated by protoc-gen-es v2.10.2 with parameter "target=ts,js_import_style=legacy_commonjs"
// @generated from file node/tag.proto (package yuhaiin.tag, edition 2023)
/* eslint-disable */

import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import { file_google_protobuf_go_features } from "@bufbuild/protobuf/wkt";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file node/tag.proto.
 */
export const file_node_tag: GenFile = /*@__PURE__*/
  fileDesc("Cg5ub2RlL3RhZy5wcm90bxILeXVoYWlpbi50YWciRgoEdGFncxILCgN0YWcYASABKAkSIwoEdHlwZRgDIAEoDjIVLnl1aGFpaW4udGFnLnRhZ190eXBlEgwKBGhhc2gYAiADKAkqIAoIdGFnX3R5cGUSCAoEbm9kZRAAEgoKBm1pcnJvchABQjZaLGdpdGh1Yi5jb20vQXN1dG9ydWZhL3l1aGFpaW4vcGtnL3Byb3Rvcy9ub2RlkgMF0j4CEANiCGVkaXRpb25zcOgH", [file_google_protobuf_go_features]);

/**
 * @generated from message yuhaiin.tag.tags
 */
export type tags = Message<"yuhaiin.tag.tags"> & {
  /**
   * @generated from field: string tag = 1;
   */
  tag: string;

  /**
   * @generated from field: yuhaiin.tag.tag_type type = 3;
   */
  type: tag_type;

  /**
   * @generated from field: repeated string hash = 2;
   */
  hash: string[];
};

/**
 * Describes the message yuhaiin.tag.tags.
 * Use `create(tagsSchema)` to create a new message.
 */
export const tagsSchema: GenMessage<tags> = /*@__PURE__*/
  messageDesc(file_node_tag, 0);

/**
 * @generated from enum yuhaiin.tag.tag_type
 */
export enum tag_type {
  /**
   * @generated from enum value: node = 0;
   */
  node = 0,

  /**
   * @generated from enum value: mirror = 1;
   */
  mirror = 1,
}

/**
 * Describes the enum yuhaiin.tag.tag_type.
 */
export const tag_typeSchema: GenEnum<tag_type> = /*@__PURE__*/
  enumDesc(file_node_tag, 0);